Update DaggerFallUI.cs
Attempted to build the fonts folder in a new location in order to be able to load the fonts in-game. With the original set up, Unity could not find /Fonts in the Assets folder. It should now copy the fonts into the user settings folder (emulated/0/android/data/com.daggerfallunityandroid.example.../files/font) and load them from there. It's a hacky solution but the best i can do for now.